refactor(toggleTheme): Change from hook to function, move to `lib/utils`
- Rename file to `toggleTheme.ts` and move to `lib/utils` directory
- Add `toggleTheme.ts` export to `index.ts`, allowing imports from `lib/utils`
- Update imports and usage
